`constrained-generators`: Simplify `HasSpec` instance for `Bool` and move more things to `NumSpec` module
