+++ libs/constrained-generators/src/Constrained/Spec/Map.hs
@@ -25,7 +25,7 @@ import Constrained.FunctionSymbol
 import Constrained.GenT
 import Constrained.Generic (Prod (..))
 import Constrained.List
-import Constrained.NumSpec (cardinality, geqSpec, leqSpec, nubOrd)
+import Constrained.NumOrd (cardinality, geqSpec, leqSpec, nubOrd)
 import Constrained.PrettyUtils
 import Constrained.Spec.Set
 import Constrained.Spec.SumProd
